+++ src/main/java/com/uiowa/fse_project/service/PatientService.java
@@ -3,5 +3,5 @@ package com.uiowa.fse_project.service;
 public interface PatientService {
    
     void selectDoctor(long patientID, long doctorID);
-    void payBill(long id, float amount);
+    void payBill(long id, long amount);
 }
